Build QScintilla2 on Windows instead of using precompiled binaries

+++ prepare.sh
@@ -214,26 +214,33 @@ fi
 
 echo -n "Do you want to use QScintilla2 editor? [y/N]: "
 read use_qsci_user
+echo -n "Use systemwide QScintilla2 library? [Y/n]: "
+read use_qsci_system
 echo ""
 if [ "$use_qsci_user" == "y" ] || [ "$use_qsci_user" == "Y" ]; then
-    echo -n "Checking for QScintilla dev files..."
-    cat > /tmp/test.cpp << EOF
+    if [ "$use_qsci_system" == "n" ] || [ "$use_qsci_system" == "N" ]; then
+        echo -n "Checking for QScintilla dev files..."
+        cat > /tmp/test.cpp << EOF
 #include <Qsci/qsciscintilla.h>
 #include <Qsci/qscilexerjavascript.h>
 #include <Qsci/qscilexerpython.h>
 int main() { QsciScintilla *s = new QsciScintilla; delete s; return 0; }
 EOF
-    $cpp_bin /tmp/test.cpp -o /tmp/test -I"$($qmake_bin -query QT_INSTALL_HEADERS)" -I"$($qmake_bin -query QT_INSTALL_HEADERS)/QtCore" -I"$($qmake_bin -query QT_INSTALL_HEADERS)/QtGui" -lqscintilla2 -lQtGui -lQtCore
-    if [ $? -eq 0 ] ; then
-        echo "ok"
-        rm /tmp/test.cpp
-        rm /tmp/test
-        GOT_QSCI=1
+        $cpp_bin /tmp/test.cpp -o /tmp/test -I"$($qmake_bin -query QT_INSTALL_HEADERS)" -I"$($qmake_bin -query QT_INSTALL_HEADERS)/QtCore" -I"$($qmake_bin -query QT_INSTALL_HEADERS)/QtGui" -lqscintilla2 -lQtGui -lQtCore
+        if [ $? -eq 0 ] ; then
+            echo "ok"
+            rm /tmp/test.cpp
+            rm /tmp/test
+            GOT_QSCI=1
+        else
+            echo "failed"
+            echo "Install qscintilla devel files!"
+            rm /tmp/test.cpp
+            exit 1
+        fi
     else
-        echo "failed"
-        echo "Install qscintilla devel files!"
-        rm /tmp/test.cpp
-        exit 1
+        GOT_QSCI=2
+        echo "Using Lorris' QScintilla2 library"
     fi
 else
     echo "Not using QScintilla"
@@ -263,6 +270,8 @@ fi
 
 if [ $GOT_QSCI -eq 1 ] ; then 
     echo "CONFIG += qsci_editor" >> config.pri
+elif [ $GOT_QSCI -eq 2 ] ; then
+    echo "CONFIG += qsci_system" >> config.pri
 fi
 
 if [ $GOT_JOYSTICK -eq 1 ] ; then
